src/requirements.txt
Flask==1.1.4
requests==2.25.1
pytest==6.2.4
couchbase==3.2.0
flask_restx==0.4.0
python-dotenv==0.19.0
bcrypt==3.2.0
six==1.15.0